# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plexo API
  description: AI agent operations platform API
  version: 0.8.0-beta.1
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Joeybuilt
    url: https://getplexo.com
servers:
  - url: /api/v1
    description: Default (relative)
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      operationId: listWorkspaces
      responses:
        '200':
          description: Workspace list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
    post:
      tags: [Workspaces]
      summary: Create workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 200
      responses:
        '201':
          description: Workspace created
  /workspaces/{id}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      operationId: getWorkspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workspace details
...
components:
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - BearerAuth: []
  - ApiKeyAuth: []